+++ examples/node_informer.rs
@@ -1,23 +1,22 @@
 #[macro_use] extern crate log;
 use kube::{
-    api::{RawApi, Informer, WatchEvent, ObjectList, Object},
+    api::{RawApi, Api, v1Event, Informer, ListParams, WatchEvent, Object},
     client::APIClient,
     config,
 };
-use k8s_openapi::api::core::v1::{
-    NodeSpec, NodeStatus,
-    Event, ListEventForAllNamespacesOptional,
-};
+use k8s_openapi::api::core::v1::{NodeSpec, NodeStatus};
 
 type Node = Object<NodeSpec, NodeStatus>;
+type Event = v1Event; // snowflake obj
 
 fn main() -> Result<(), failure::Error> {
-    std::env::set_var("RUST_LOG", "info,kube=trace");
+    std::env::set_var("RUST_LOG", "info,node_informer=debug,kube=debug");
     env_logger::init();
     let config = config::load_kube_config().expect("failed to load kubeconfig");
     let client = APIClient::new(config);
 
     let nodes = RawApi::v1Node();
+    let events = Api::v1Event(client.clone());
     let ni = Informer::raw(client.clone(), nodes)
         .labels("role=worker")
         .init()?;
@@ -25,15 +24,15 @@ fn main() -> Result<(), failure::Error> {
     loop {
         ni.poll()?;
 
-        while let Some(event) = ni.pop() {
-            handle_nodes(&client, event)?;
+        while let Some(ne) = ni.pop() {
+            handle_nodes(&events, ne)?;
         }
     }
 }
 
 // This function lets the app handle an event from kube
-fn handle_nodes(client: &APIClient, ev: WatchEvent<Node>) -> Result<(), failure::Error> {
-    match ev {
+fn handle_nodes(events: &Api<Event>, ne: WatchEvent<Node>) -> Result<(), failure::Error> {
+    match ne {
         WatchEvent::Added(o) => {
             info!("New Node: {}", o.spec.provider_id.unwrap());
         },
@@ -47,16 +46,15 @@ fn handle_nodes(client: &APIClient, ev: WatchEvent<Node>) -> Result<(), failure:
                     (c.status == "False" &&  c.type_ == "Ready")
                 }).map(|c| c.message).collect::<Vec<_>>(); // failed statuses
                 warn!("Unschedulable Node: {}, ({:?})", o.metadata.name, failed);
-                // Separate API call with client to find events related to this node
+                // Find events related to this node
                 let sel = format!("involvedObject.kind=Node,involvedObject.name={}", o.metadata.name);
-                let opts = ListEventForAllNamespacesOptional {
-                    field_selector: Some(&sel),
+                let opts = ListParams {
+                    field_selector: Some(sel),
                     ..Default::default()
                 };
-                let req = Event::list_event_for_all_namespaces(opts)?.0;
-                let res = client.request::<ObjectList<Event>>(req)?;
-                for e in res.items {
-                    warn!("Node event: {:?}", e);
+                let evlist = events.list(&opts)?;
+                for e in evlist.items {
+                    warn!("Node event: {:?}", serde_json::to_string_pretty(&e)?);
                 }
             } else {
                 // Turn up logging above to see