+++ src/api/mod.rs
@@ -1,7 +1,7 @@
 //! API helpers
 
 /// Empty struct for when no Spec is required
-#[derive(Clone, Deserialize, Default)]
+#[derive(Clone, Deserialize, Serialize, Default)]
 pub struct Void {}
 
 mod reflector;
@@ -54,4 +54,5 @@ pub use self::metadata::{
     ObjectMeta,
     TypeMeta,
     Initializers,
+    OwnerReference,
 };